gis/2-flask-solucion_gis_completo/requirements.txt

Flask
psycopg2-binary
